remove legacy list component
